Remove deprecated tensorflow_io.json and reroute to tfio.IODataset.from_json
tfio.IODataset.from_json is already deprecated some time ago,
this move is repointing to tfio.IODataset.from_json
