+++ tensorflow_io/core/BUILD
@@ -375,14 +375,28 @@ cc_library(
         "ops/grpc_ops.cc",
     ],
     copts = tf_io_copts(),
-    includes = [
-        ".",
-    ],
     linkstatic = True,
     deps = [
         "//tensorflow_io/core:dataset_ops",
         "//tensorflow_io/core/grpc:endpoint_cc_grpc",
         "@com_github_grpc_grpc//:grpc++",
+    ],
+    alwayslink = 1,
+)
+
+cc_library(
+    name = "json_ops",
+    srcs = [
+        "kernels/json_kernels.cc",
+        "ops/json_ops.cc",
+    ],
+    copts = tf_io_copts(),
+    linkstatic = True,
+    deps = [
+        "//tensorflow_io/arrow:arrow_ops",
+        "//tensorflow_io/core:dataset_ops",
+        "@arrow",
+        "@jsoncpp_git//:jsoncpp",
     ],
     alwayslink = 1,
 )
@@ -404,6 +418,7 @@ cc_binary(
         "//tensorflow_io/core:grpc_ops",
         "//tensorflow_io/core:hdf5_ops",
         "//tensorflow_io/core:image_ops",
+        "//tensorflow_io/core:json_ops",
         "//tensorflow_io/core:kinesis_ops",
         "//tensorflow_io/core:lmdb_ops",
         "//tensorflow_io/core:parquet_ops",
@@ -414,7 +429,6 @@ cc_binary(
         "//tensorflow_io/core/oss:oss_ops",
         "//tensorflow_io/gcs:gcs_config_ops",
         "//tensorflow_io/ignite:ignite_ops",
-        "//tensorflow_io/json:json_ops",
         "//tensorflow_io/kafka:kafka_ops",
         "@libarchive",
         "@local_config_tf//:libtensorflow_framework",
