--- tensorflow_io/json/__init__.py
@@ -1,38 +1,0 @@
-# Copyright 2018 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""JSONDataset.
-
-@@JSONDataset
-@@list_json_columns
-@@read_json
-"""
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from tensorflow.python.util.all_util import remove_undocumented
-
-from tensorflow_io.json.python.ops.json_ops import JSONDataset
-from tensorflow_io.json.python.ops.json_ops import list_json_columns
-from tensorflow_io.json.python.ops.json_ops import read_json
-
-_allowed_symbols = [
-    "JSONDataset",
-    "list_json_columns",
-    "read_json",
-]
-
-remove_undocumented(__name__, allowed_exception_list=_allowed_symbols)
