+++ tests/test_json_eager.py
@@ -21,10 +21,7 @@ from __future__ import print_function
 import os
 
 import tensorflow as tf
-if not (hasattr(tf, "version") and tf.version.VERSION.startswith("2.")):
-  tf.compat.v1.enable_eager_execution()
-import tensorflow_io as tfio  # pylint: disable=wrong-import-position
-import tensorflow_io.json as json_io  # pylint: disable=wrong-import-position
+import tensorflow_io as tfio
 
 def test_io_tensor_json_recods_mode():
   """Test case for tfio.IOTensor.from_json."""
@@ -63,13 +60,9 @@ def test_io_tensor_json_recods_mode():
     assert v_y[1] == integer_label[i].numpy()
 
   feature_dataset = tfio.IODataset.from_json(feature_filename, mode='records')
-
   label_dataset = tfio.IODataset.from_json(label_filename, mode='records')
 
-  dataset = tf.data.Dataset.zip((
-      feature_dataset,
-      label_dataset
-  ))
+  dataset = tf.data.Dataset.zip((feature_dataset, label_dataset))
 
   i = 0
   for (j_x, j_y) in dataset:
@@ -163,21 +156,11 @@ def test_json_dataset():
       "label.json")
   label_filename = "file://" + label_filename
 
-  feature_cols = json_io.list_json_columns(feature_filename)
-  assert feature_cols["floatfeature"].dtype == tf.float64
-  assert feature_cols["integerfeature"].dtype == tf.int64
+  feature_dataset = tfio.IODataset.from_json(
+      feature_filename, ["floatfeature", "integerfeature"], mode="records")
 
-  label_cols = json_io.list_json_columns(label_filename)
-  assert label_cols["floatlabel"].dtype == tf.float64
-  assert label_cols["integerlabel"].dtype == tf.int64
-
-  feature_dataset = json_io.JSONDataset(
-      feature_filename, ["floatfeature", "integerfeature"]
-  ).apply(tf.data.experimental.unbatch())
-
-  label_dataset = json_io.JSONDataset(
-      label_filename, ["floatlabel", "integerlabel"]
-  ).apply(tf.data.experimental.unbatch())
+  label_dataset = tfio.IODataset.from_json(
+      label_filename, ["floatlabel", "integerlabel"], mode="records")
 
 
   dataset = tf.data.Dataset.zip((
@@ -212,14 +195,18 @@ def test_json_keras():
   feature_cols = ["sepalLength", "sepalWidth", "petalLength", "petalWidth"]
   label_cols = ["species"]
 
-  feature_dataset = json_io.JSONDataset(feature_filename, feature_cols)
+  feature_dataset = tfio.IODataset.from_json(
+      feature_filename, feature_cols, mode="records")
 
-  label_dataset = json_io.JSONDataset(label_filename, label_cols)
+  label_dataset = tfio.IODataset.from_json(
+      label_filename, label_cols, mode="records")
 
   dataset = tf.data.Dataset.zip((
       feature_dataset,
       label_dataset
   ))
+  dataset = dataset.batch(1)
+
   def pack_features_vector(features, labels):
     """Pack the features into a single array."""
     features = tf.stack(list(features), axis=1)
